+++ src/ghrelease/utils.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+import os
 import mimetypes
 
 from .log import t
@@ -50,10 +51,10 @@ def get_assets(release):
     return {asset.name: asset for asset in release.iter_assets()}
 
 def upload_assets(release, assets):
-
     existing_assets = get_assets(release)
 
-    for filename in assets:
+    for asset_path in assets:
+        filename = os.path.basename(asset_path)
         existing = existing_assets.get(filename)
         if existing:
             error(12, "Asset `%s` already exists: `%s/%s`" % (filename, release.html_url, existing.name))